+++ src/index.ts
@@ -5,10 +5,12 @@ export * from "./constant/column.enum";
 export * from "./constant/options.const";
 
 export * from "./decorator/Checks";
+export * from "./decorator/IndexColumns";
 export * from "./decorator/UniqueColumns";
 
 export * from "./util/crypt-sha1";
 export * from "./util/safe-constraint";
+export * from "./util/safe-index";
 export * from "./util/safe-unique";
 export * from "./util/strip-public";
 export * from "./util/strip-schema";